pandas
Pandas is a Python library used for working with data sets. 
It has functions for analyzing, cleaning, exploring, and manipulating data. 
The name "Pandas" has a reference to both "Panel Data", and "Python Data Analysis" and was created by Wes McKinney in 2008.
If you have Python and PIP already installed on a system, then installation of Pandas is very easy.

Install it using this command:
pip install pandas

Pandas is usually imported under the pd alias.

Now the Pandas package can be referred to as pd instead of pandas.

pandas is a data manipulation package in Python for tabular data. 
That is, data in the form of rows and columns, also known as DataFrames. 
Intuitively, you can think of a DataFrame as an Excel sheet.